Consensus Algorithms: Move template functions into .h file.
