+++ source/base/CMakeLists.txt
@@ -50,7 +50,6 @@ SET(_unity_include_src
   hdf5.cc
   mpi.cc
   mpi_compute_index_owner_internal.cc
-  mpi_consensus_algorithms.cc
   mpi_noncontiguous_partitioner.cc
   mpi_remote_point_evaluation.cc
   mu_parser_internal.cc
