+++ source/base/mpi_compute_index_owner_internal.cc
@@ -18,6 +18,8 @@
 #include <deal.II/base/mpi.h>
 #include <deal.II/base/mpi_compute_index_owner_internal.h>
 
+#include <boost/serialization/utility.hpp>
+
 DEAL_II_NAMESPACE_OPEN
 
 namespace Utilities
